requirements.txt
# Support library for versioning
importlib-metadata

# Framework and libraries
flask
python-dotenv

# Git hooks
pre-commit

# More production-ready web server
#gunicorn
